pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dms-frontend"
version = "0.1.0"
description = "Data Moving Service frontend API"
authors = [{name = "DMS Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.26",
    "redis>=5.0",
    "pydantic-settings>=2.2",
    "typer>=0.9",
    "python-json-logger>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "httpx>=0.26",
]

[project.scripts]
dms-frontend = "cli.main:main"

[tool.setuptools]
packages = ["cli", "app"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
